testin gamestate

+++ server/server_structures.h
@@ -4,11 +4,10 @@
 
 #ifndef STRUCTURES_H
 #define STRUCTURES_H
-
 #define MAX_BULLETS 128
 struct  coordinate {
-    int16_t x;
-    int16_t y;
+    float x;
+    float y;
 };
 
 struct bullet {
